[ADD] düzenlenecek malzeme bilgilerini formatlama
