+++ src/main/resources/templates/statuonay.html
@@ -13,17 +13,7 @@
         <title>Malzeme Zimmetleme</title>
     </head>
 <body>
-<!--
-@TODO
-tüm malzemeleri listele
-altına bir input form koy
-inputun altına bir submint button
-tabloyu seçlebilir yap
-controllera geç
 
-
-
--->
 <div th:insert="header:: menu"></div>
 <div class="container">
     <h2>Malzemeler</h2>
